Reparation des fonctions de Delete (le cascadeOnDelete marche pas donc je fais ça au niveau des modeles  )

+++ app/Models/Evenement.php
@@ -61,7 +61,14 @@ class Evenement extends Model
         if($res==true)
         {
             $relations=$this->image; // here get the relation data
-            $relations->delete();// delete Here
+            if($relations){
+
+                $relations->delete();// delete Here
+            }
+
+             $this->creneaus->each(function(Creneau $creneau){
+                 $creneau->delete();
+             });
 
         }
     }
